Change data type from Double to Float.

+++ app/src/main/java/de/example/crudapp/data/networking/FlaskProductRestApi.kt
@@ -15,7 +15,7 @@ interface FlaskProductRestApi {
     suspend fun postProduct(
         @Field("name") name: String,
         @Field("description") description: String,
-        @Field("price") price: Double,
+        @Field("price") price: Float,
         @Field("qty") qty: Int
     ): Response<Product>
 
@@ -28,7 +28,7 @@ interface FlaskProductRestApi {
         @Path("id") id: Int,
         @Field("name") name: String,
         @Field("description") description: String,
-        @Field("price") price: Double,
+        @Field("price") price: Float,
         @Field("qty") qty: Int
     ): Response<Product>
